+++ docs/versionhistory.rst
@@ -3,6 +3,11 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- Fixed ``RuntimeError`` on asyncio when a ``CancelledError`` is raised from a task spawned through
+  a ``BlockingPortal`` (`#357 <https://github.com/agronholm/anyio/issues/357>`_)
+
 **3.3.0**
 
 - Added asynchronous ``Path`` class